余分なファイル削除
