+++ PetSNS/Models/user/Owner.swift
@@ -14,3 +14,35 @@ struct Owner {
     var allFollow: [Follow]
     var allFollower: [Follower]
 }
+
+struct Pet {
+    var id: String
+    var image: UIImage = #imageLiteral(resourceName: "スクリーンショット 2021-09-16 1.09.47.png")
+    var name: String
+    var introduction: String = ""
+    var posts: [Post] = []
+    var followId: String
+}
+
+struct Follow {
+    var id: String
+    var follow: [String] = []
+}
+
+struct Follower {
+    var id: String
+    var follower: [String] = []
+}
+
+struct Post {
+    var id: String
+    var image: UIImage
+    var text: String = ""
+    var likedUsers: [String] = []
+    var comments: [Comment] = []
+}
+
+struct Comment {
+    let user: String
+    let comment: String
+}
